requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
google-generativeai>=0.4.0
pytz>=2024.1